# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RGUKT M.Tech Results Portal API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface StudentLoginInput {
  studentId: string;
  password: string;
}

export interface AdminLoginInput {
  username: string;
  password: string;
}

export interface StudentSession {
  studentId: string;
  name: string;
  role: string;
}

export interface AdminSession {
  username: string;
  role: string;
}

export interface SessionInfo {
  role: string;
  id: string;
  name: string;
}

export interface Student {
  id: number;
  studentId: string;
  name: string;
  branch: string;
  batch: string;
  createdAt: string;
}

export interface StudentInput {
  studentId: string;
  name: string;
  branch: string;
  batch: string;
  password?: string;
}

export interface StudentUpdate {
  name?: string;
  branch?: string;
  batch?: string;
}

export interface Result {
  id: number;
  studentId: string;
  semester: number;
  subjectCode: string;
  subjectName: string;
  credits: number;
  grade: string;
  gradePoint: number;
  createdAt: string;
}

export interface SemesterResult {
  semester: number;
  sgpa: number;
  totalCredits: number;
  results: Result[];
}

export interface StudentResultsSummary {
  student: Student;
  semesters: SemesterResult[];
  cgpa: number;
}

export interface PdfUploadInput {
  filename: string;
}

export interface PdfUploadResult {
  message: string;
  extracted: number;
  inserted: number;
  updated: number;
  studentsAutoCreated: number;
  errors: number;
  uploadId: number;
}

export interface PdfUpload {
  id: number;
  filename: string;
  uploadedAt: string;
  recordsExtracted: number;
  recordsInserted: number;
  studentsCreated: number;
}

export interface AdminStats {
  totalStudents: number;
  totalResults: number;
  totalUploads: number;
  branches: string[];
  recentUploads: PdfUpload[];
}

export type ListStudentsParams = {
search?: string;
branch?: string;
};

export type ListResultsParams = {
studentId?: string;
semester?: number;
};