# lib/api-zod/src/generated/types/adminSession.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RGUKT M.Tech Results Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminSession {
  username: string;
  role: string;
}