# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RGUKT M.Tech Results Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { PdfUpload } from './pdfUpload';

export interface AdminStats {
  totalStudents: number;
  totalResults: number;
  totalUploads: number;
  branches: string[];
  recentUploads: PdfUpload[];
}