# lib/api-zod/src/generated/types/pdfUploadResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RGUKT M.Tech Results Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface PdfUploadResult {
  message: string;
  extracted: number;
  inserted: number;
  updated: number;
  studentsAutoCreated: number;
  errors: number;
  uploadId: number;
}