# lib/api-zod/src/generated/types/semesterResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RGUKT M.Tech Results Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { Result } from './result';

export interface SemesterResult {
  semester: number;
  sgpa: number;
  totalCredits: number;
  results: Result[];
}