# lib/api-zod/src/generated/types/sessionInfo.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RGUKT M.Tech Results Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface SessionInfo {
  role: string;
  id: string;
  name: string;
}