# lib/api-zod/src/generated/types/studentLoginInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RGUKT M.Tech Results Portal API
 * OpenAPI spec version: 0.1.0
 */

export interface StudentLoginInput {
  studentId: string;
  password: string;
}