# lib/api-zod/src/generated/types/studentResultsSummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * RGUKT M.Tech Results Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { SemesterResult } from './semesterResult';
import type { Student } from './student';

export interface StudentResultsSummary {
  student: Student;
  semesters: SemesterResult[];
  cgpa: number;
}